[refactor] expression 클래스 분리
